✨ Create search filter on lessons page
