+++ components/Lessons/Lessons.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import styles from "./Lessons.module.css";
 import ListItems from "../ListItems/ListItem";
 import LessonsHead from "../LessonsHead/LessonsHead";
@@ -6,16 +6,28 @@ import ButtonGroup from "../ButtonGroup/ButtonGroup";
 import SearchInput from "../SearchInput/SearchInput";
 
 function Lessons() {
+  const [searchValue, setSearchValue] = useState("");
+  const [tabs, setTabs] = useState(null);
+
+  useEffect(() => {
+    fetch(`/api/tabs${searchValue ? `?search=${searchValue}` : ""}`)
+      .then((result) => result.json())
+      .then(setTabs);
+  }, [searchValue]);
+
   return (
     <div className={styles.container}>
       <LessonsHead label={"Beginner"} link={"/beginner"} />
 
-      <SearchInput />
+      <SearchInput search={searchValue} onSearch={setSearchValue} />
+
       <div>
-        <ListItems
-          title={"Stairway To Heaven - Led Zeppelin"}
-          checkBox={false}
-        />
+        {tabs &&
+          tabs.map((tab) => {
+            return (
+              <ListItems key={tab._id} title={tab.title} checkBox={false} />
+            );
+          })}
       </div>
       <ButtonGroup />
     </div>
